Add utils functions & update comments
